migrate push unlock and push appium
