+++ lib/devices/android/adb.js
@@ -889,53 +889,6 @@ ADB.prototype.getLogcatLogs = function() {
   return this.logcat.getLogs();
 };
 
-ADB.prototype.pushAppium = function(cb) {
-  this.debug("Pushing appium bootstrap to device...");
-  var binPath = path.resolve(__dirname, "..", "..", "..", "build",
-      "android_bootstrap", "AppiumBootstrap.jar");
-  fs.stat(binPath, function(err) {
-    if (err) {
-      cb(new Error("Could not find AppiumBootstrap.jar; please run " +
-                   "'grunt buildAndroidBootstrap'"));
-    } else {
-      var remotePath = "/data/local/tmp";
-      var cmd = this.adbCmd + ' push "' + binPath + '" ' + remotePath;
-      exec(cmd, { maxBuffer: 524288 }, function(err) {
-        if (err) {
-          logger.error(err);
-          cb(err);
-        } else {
-          cb(null);
-        }
-      }.bind(this));
-    }
-  }.bind(this));
-};
-
-ADB.prototype.pushUnlock = function(cb) {
-  // TODO: calling `adb install` may not be necessary if its already there.
-  // can we check if app exists first? may speed this up.
-  this.debug("Pushing unlock helper app to device...");
-  var unlockPath = path.resolve(__dirname, "..", "..", "..", "build",
-      "unlock_apk", "unlock_apk-debug.apk");
-  fs.stat(unlockPath, function(err) {
-    if (err) {
-      cb(new Error("Could not find unlock.apk; please run " +
-                   "'reset.sh --android' to build it."));
-    } else {
-      var cmd = this.adbCmd + ' install "' + unlockPath + '"';
-      exec(cmd, { maxBuffer: 524288 }, function(err) {
-        if (err) {
-          logger.error(err);
-          cb(err);
-        } else {
-          cb(null);
-        }
-      }.bind(this));
-    }
-  }.bind(this));
-};
-
 ADB.prototype.startApp = function(cb) {
   logger.info("Starting app");
   this.requireDeviceId();
@@ -1102,29 +1055,21 @@ ADB.prototype.uninstallApk = function(pkg, cb) {
   });
 };
 
-ADB.prototype.installApk = function(apk, cb) {
-  var cmd = this.adbCmd + ' install -r "' + apk + '"';
-  logger.info("Installing " + apk);
-  exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
-    if (err) {
-      logger.error(err);
-      cb(err);
-    } else {
-      // Useful for debugging.
-      logger.debug(stdout);
-      cb(null);
-    }
-  });
-};
-
-
 ADB.prototype.installRemote = function(remoteApk, cb) {
   var cmd = 'pm install -r ' + remoteApk;
   this.shell(cmd, cb);
 };
 
-ADB.prototype.install = function(apk, cb) {
-  var cmd = 'install -r ' + apk;
+ADB.prototype.install = function(apk, replace, cb) {
+  if (typeof replace === "function") {
+    cb = replace;
+    replace = true;
+  }
+  var cmd = 'install ';
+  if (replace) {
+    cmd += '-r ';
+  }
+  cmd += apk;
   this.exec(cmd, cb);
 };
