+++ lib/devices/android/android-common.js
@@ -2,6 +2,7 @@
 var logger = require('../../server/logger.js').get('appium')
   , _ = require('underscore')
   , fs = require('fs')
+  , path = require('path')
   , md5 = require('MD5')
   , async = require('async');
 
@@ -171,5 +172,20 @@ androidCommon.removeTempApks = function(exceptMd5s, cb) {
 androidCommon.forwardPort = function(cb) {
   this.adb.forwardPort(this.opts.systemPort, this.opts.devicePort, cb);
 };
+
+androidCommon.pushUnlock = function(cb) {
+  logger.debug("Pushing unlock helper app to device...");
+  var unlockPath = path.resolve(__dirname, "..", "..", "..", "build",
+      "unlock_apk", "unlock_apk-debug.apk");
+  fs.stat(unlockPath, function(err) {
+    if (err) {
+      cb(new Error("Could not find unlock.apk; please run " +
+                   "'reset.sh --android' to build it."));
+    } else {
+      this.adb.install(unlockPath, false, cb);
+    }
+  }.bind(this));
+};
+
 
 module.exports = androidCommon;
