+++ lib/devices/android/android.js
@@ -207,6 +207,21 @@ Android.prototype.pushStrings = function(cb) {
   }
 };
 
+Android.prototype.pushAppium = function(cb) {
+  logger.debug("Pushing appium bootstrap to device...");
+  var binPath = path.resolve(__dirname, "..", "..", "..", "build",
+      "android_bootstrap", "AppiumBootstrap.jar");
+  fs.stat(binPath, function(err) {
+    if (err) {
+      cb(new Error("Could not find AppiumBootstrap.jar; please run " +
+                   "'grunt buildAndroidBootstrap'"));
+    } else {
+      this.adb.push(binPath, this.remoteTempDir(), cb);
+    }
+  }.bind(this));
+};
+
+
 Android.prototype.timeoutWaitingForCommand = function() {
   logger.info("Didn't get a new command in " + (this.commandTimeoutMs / 1000) +
               " secs, shutting down...");
@@ -230,7 +245,7 @@ Android.prototype.uninstallApp = function(cb) {
   }.bind(this);
 
   if (this.skipUninstall) {
-    this.debug("Not uninstalling app since server not started with " +
+    logger.debug("Not uninstalling app since server not started with " +
                "--full-reset");
     cb();
   } else {
